add: diary controller, diary service

+++ src/main/java/org/dongguk/diary/repository/DiaryRepository.java
@@ -1,12 +1,12 @@
 package org.dongguk.diary.repository;
 
-import org.dongguk.diary.entity.Diary;
+import org.dongguk.diary.domain.Diary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
-    Optional<Diary> findByDid(Integer did);
+public interface DiaryRepository extends JpaRepository<Diary, Long> {
+    Optional<Diary> findByDid(Long did);
 }
